// frontend/src/shared/types/MaterialType.ts
/** Categoria de material/produs a unui element de cumpărat. */
export enum MaterialType {
  Gresie = "Gresie",
  Faianta = "Faianță",
  Plinta = "Plintă",
  Parchet = "Parchet",
  Vopsea = "Vopsea",
  Tapet = "Tapet",
  GlafFereastra = "Glaf Fereastră",
  Sanitare = "Sanitare",
  Mobila = "Mobilă",
  Electrocasnice = "Electrocasnice",
  CorpuriIluminat = "Corpuri de iluminat",
  Altele = "Altele",
}